requirements.txt
Flask
sentence-transformers
numpy
scikit-learn
python-dotenv
